7.0

+++ cssedit.php
@@ -34,6 +34,8 @@ require ("./data/include/errorhandler.inc.php");
 require ("./data/include/siteinfo.inc.php");
 // define page specific vars -----------------
 // define page specific funcs ----------------
+
+// page actions
 if (isset($_POST['op']) && $_POST['op']=="save" && trim($_POST['file'])!='') {
     $timestamp = time();
     if (intval($_POST['id'])>0) {
@@ -70,14 +72,14 @@ if (isset($_POST['op']) && $_POST['op']=="save" && trim($_POST['file'])!='') {
 }
 else if (isset($_POST['op']) && $_POST['op']=="savefolder" && isset($_POST['id']) && intval($_POST['id'])>0) {
     // update css folder
-    $stylesheet = NULL;
+    $stylesheet = null;
     if (array_key_exists('stylesheet', $_POST)) {
         $stylesheet = serialize($_POST['stylesheet']);
     }
     $sql = "UPDATE `stylesheets` SET 
         `stylesheet` = '".escapeSQL($stylesheet)."',
         `describ` = '".escapeSQL(trim($_POST['describfolder']))."',
-        `lastchange` = '".time()."'
+        `lastchange` = ".time()."
         WHERE `id` = ".intval($_POST['id']);
     $res = doSQL($sql);
     if ($res['aff']==1) {
@@ -126,31 +128,35 @@ else if (isset($_FILES['uploadfolder']) && trim($_FILES['uploadfolder']['tmp_nam
                 // extract all files
                 $phar = new PharData(cleanPath(DOCUMENT_ROOT.DIRECTORY_SEPARATOR.$filename));
                 $phar->extractTo(cleanPath(DOCUMENT_ROOT.DIRECTORY_SEPARATOR.$foldername));
-                $emptyextract = clearFolder($foldername, array('.css'));
+                $emptyextract = clearFolder($foldername, array('.css','.map','.png','.gif','.svg','.jpg','.otf','.woff','.woff2','.eot','.ttf','.md'));
                 if ($emptyextract===true) {
                     deleteFolder($foldername, false);
                     addWSPMsg('errormsg', returnIntLang('folder upload was empty or had false contents', false));
                 } else {
-                    copyFolder($foldername, cleanPath('/media/layout/'.$scriptname.'/'));
-                    deleteFolder($foldername, false);
-                    deleteFile(cleanPath($filename));
-                    // check for existing folder
-                    $sql = "SELECT `id` FROM `stylesheets` WHERE `cfolder` = '".escapeSQL(urltext($scriptname))."'";
-                    $res = doResultSQL($sql);
-                    if ($res===false) {
-                        // insert some data to database for a new entry
-                        $sql = "INSERT INTO `stylesheets` SET `file` = '', `cfolder` = '".escapeSQL(urltext($scriptname))."', `describ` = '".escapeSQL(urltext($scriptname)." ".returnIntLang('str folder', false))."', `lastchange` = ".time();
-                        $res = doSQL($sql);
-                        if ($res['inf']>0) {
-                            addWSPMsg('resultmsg', returnIntLang('folder upload done', false));
+                    $copyfolder = copyFolder($foldername, cleanPath('/media/layout/'.$scriptname.'/'));
+                    if ($copyfolder) {
+                        deleteFolder($foldername, false);
+                        deleteFile(cleanPath($filename));
+                        // check for existing folder
+                        $sql = "SELECT `id` FROM `stylesheets` WHERE `cfolder` = '".escapeSQL($scriptname)."'";
+                        $res = doResultSQL($sql);
+                        if ($res===false) {
+                            // insert some data to database for a new entry
+                            $sql = "INSERT INTO `stylesheets` SET `file` = '', `cfolder` = '".escapeSQL($scriptname)."', `describ` = '".escapeSQL($scriptname." ".returnIntLang('str folder', false))."', `lastchange` = ".time();
+                            $res = doSQL($sql);
+                            if ($res['inf']>0) {
+                                addWSPMsg('resultmsg', returnIntLang('css folder upload done', false));
+                            }
                         }
-                    }
-                    else {
-                        $sql = "UPDATE `stylesheets` SET `lastchange` = ".time()." WHERE `cfolder` = '".escapeSQL(urltext($scriptname))."'";
-                        $res = doSQL($sql);
-                        if ($res['aff']>0) {
-                            addWSPMsg('resultmsg', returnIntLang('folder upload update done', false));
+                        else {
+                            $sql = "UPDATE `stylesheets` SET `lastchange` = ".time()." WHERE `cfolder` = '".escapeSQL($scriptname)."'";
+                            $res = doSQL($sql);
+                            if ($res['aff']>0) {
+                                addWSPMsg('resultmsg', returnIntLang('css folder upload update done', false));
+                            }
                         }
+                    } else {
+                        addWSPMsg('errormsg', returnIntLang('css folder upload error uploading file', false));
                     }
                 }
             } 
@@ -226,7 +232,7 @@ $browserarray = array(
 
 // run folder for files
 if (is_dir(cleanPath(DOCUMENT_ROOT.DIRECTORY_SEPARATOR."media".DIRECTORY_SEPARATOR."layout".DIRECTORY_SEPARATOR))) {
-    $scanfiles = scanfiles(cleanPath(DIRECTORY_SEPARATOR."media".DIRECTORY_SEPARATOR."layout".DIRECTORY_SEPARATOR));
+    $scanfiles = scanfiles(cleanPath(DIRECTORY_SEPARATOR."media".DIRECTORY_SEPARATOR."layout".DIRECTORY_SEPARATOR), SCANDIR_SORT_ASCENDING , false , array('css') );
 } else {
     $created = createFolder(DIRECTORY_SEPARATOR."media".DIRECTORY_SEPARATOR."layout".DIRECTORY_SEPARATOR);
     if ($created===true) {
@@ -280,9 +286,7 @@ require("./data/include/sidebar.inc.php");
             $syscssfiles = array();
             if (is_array($css_res) && count($css_res)>0) {
                 foreach ($css_res AS $ck => $cv) {
-                    if (in_array(trim($cv).".css", $foundcssfiles)) {
-                        $syscssfiles[] = trim($cv).".css";
-                    }
+                    $syscssfiles[] = trim($cv).".css";
                 }
             }
